backend/getRecipesAPI/requirements.txt
Django==3.1
djangorestframework==3.11.1
numpy==1.18.4
scikit-learn==0.24.2
joblib==0.16.0
pandas